Domasna2/architecture/storage_service/requirements.txt
flask
flask-sqlalchemy
flask-migrate
requests
beautifulsoup4
lxml
pandas
mysql-connector-python
python-dateutil
pymysql
